Created actions array on socket io API
