--- src/pages/api/socketio.js
@@ -1,29 +1,0 @@
-import { Server } from 'socket.io';
-
-const ioHandler = (req, res) => {
-  if (!res.socket.server.io) {
-    console.log('*First use, starting socket.io');
-
-    const io = new Server(res.socket.server);
-
-    io.on('connection', (socket) => {
-      socket.on('moveList', (msg) => {
-        socket.broadcast.emit('moveList', msg);
-      });
-      socket.broadcast.emit('a user connected');
-    });
-
-    res.socket.server.io = io;
-  } else {
-    console.log('socket.io already running');
-  }
-  res.end();
-};
-
-export const config = {
-  api: {
-    bodyParser: false,
-  },
-};
-
-export default ioHandler;
